+++ src/util/serde/mod.rs
@@ -1,4 +1,5 @@
 //! Utils and helpers for serialization/deserialization with [`serde`].
 
+pub mod seconds;
 pub mod timelike;
 pub mod timezone;